~5x faster as TokenStream is now a dlinked list
A little bit harder to maintain, but at least the time
spent on insert and removal operations is now constant :)

Function calls are the bottleneck now. Not much can be done
unless compilation of the macro/parser stack into a single
function call with `gotos`. But this would require macro
declarations to be allowed only at the top of the file and
will probably kill any possibility to have "ifdefs" in the
future... food for thoughts.