finally solving the infinite macro recursion problem the right way!
cc #2

Each token and macro expansion cycle now owns a "blue context". A blue
context is an include only mask responsible to hold a list of unique macro
identifiers.

On tokens, the blue context holds a list of macros already applyed
to the subject token.

On expansion cycles, the blue context holds a list of macros already
expanded on the current iteration. Each cycle drops the old context and
boots a new one.

After being matched and before an expansion, a macro must cancel itself
whenever the intersection of the current expansion's blue context has a non
empty intersection with the blue context of one of the matched tokens.

After an expansion, a macro must include itself on the current expasion's
blue context.

TODO:

Add a way to force a macro to be recursive. Eg: the following use case
with homoiconicity might be desirable:

```
macro :recursive {
    (T_STRING·A ···rest) // matches a lisp form
} >> {
    T_STRING·A(···rest)
}

(sum 1 (multiply 2 3))

?>
```