refactor `layer()` parser and fixes old known heuristic bugs.
The layer() matcher had known bugs regarding matching pair heuristics
which caused source like `{([)}]` to match as false positive :P

Yea... it was lousy, but it was good enough for the job and the engine would
catch the syntax error later anyway. The new behavior is:

{ ( [ ) ] }
      ^ syntax error unexpected ')' expected ']'

This means that all layer matching will be quite strict from now on.

This also makes `layer()` a tad bit slower. A better approach could be mapping
closing/opening pairs of tokens ahead of time, when building the token stream,
and make layer matching very close to O(1). Routine could be like.:

```
class TokenStream {
...
    function mapPairs() {
        stack = [];
        foreach_node_on_token_stream(current) {
            switch (current.token.type) {
                case '{':
                case '(':
                case '[':
                T_CURLY_OPEN:
                T_DOLLAR_OPEN_CURLY_BRACES:
                    $stack[] = current;
                    break;

                case '}':
                case ')':
                case ']':
                    if (start = array_pop(stack)){
                        start.meta['pair'] = current;
                        current.meta['pair'] = start;
                    }
		    else {
			throw new \Exception('Unbalanced token stream');
		    }
                    break;
            }
        }

        if (stacklength > 0)
            throw new \Exception('Unbalanced token stream');
    }
...
}
```

This would come with a natural way of blocking unbalanced pair of tokens
(which is a good thing) but also with a general performance penalty.

Matching a layer would become as easy as:

```
if (current.token.type === '{') {
    current = current.next;
    end = current.meta['pair'];
    layer = [];
    while (current !== end)	{
        layer[] = current.token;
        current = current.next;
    }
}
```

But since macros with `{···match}`, `(···match)` or `[···match]` as entry
point are the only ones that could be significatively benefited with such
"ahead of time token pair mapping" optimization - and these kind if macros
are not very likely to happen - let's leave this for later consideration.