remove special case for macros with empty expansion.
This means that	the engine won't try to	elide the excessive empty space
left behind when a macro expands to nothing. This relates to code formatting
and is not YAY responsibility. Why did I edge case in the first place? :wonders:

A pretty printer should	be used after the expansion from now on, if beauty is desired :)